Master program to use
use cipher_program to run the secret messages program. Please note, when you create an encrypted message while the program is running, it stores the message to a global list. When decrypting a message, if the message is in the global list, a user will have to provide a PIN to decrypt a message.